+++ tests/conftest.py
@@ -7,110 +7,13 @@ used across all test modules.
 
 import asyncio
 import logging
-from unittest.mock import Mock, patch
+import os
+from unittest.mock import patch
 
 import pytest
 
-# Note: For now, we'll use the minimal versions from test_helpers
-# to avoid importing the full devops_agent with its dependencies
-from tests.fixtures.test_helpers import (
-    MetricsCollector,
-    MockCallbackContext,
-    MockInvocationContext,
-    MockLlmRequest,
-    MockLlmResponse,
-    MockTool,
-    MockToolContext,
-    StateManager,
-    create_populated_state_manager,
-    create_sample_legacy_state,
-)
-
 # Configure logging for tests
 logging.basicConfig(level=logging.DEBUG)
-
-
-@pytest.fixture
-def state_manager():
-    """Provide a fresh StateManager instance for each test."""
-    return StateManager()
-
-
-@pytest.fixture
-def populated_state_manager():
-    """Provide a StateManager with sample data for testing."""
-    return create_populated_state_manager()
-
-
-@pytest.fixture
-def sample_legacy_state():
-    """Provide sample legacy state format for testing."""
-    return create_sample_legacy_state()
-
-
-@pytest.fixture
-def mock_invocation_context():
-    """Provide a mock invocation context for testing."""
-    return MockInvocationContext()
-
-
-@pytest.fixture
-def mock_callback_context():
-    """Provide a mock callback context for testing."""
-    return MockCallbackContext()
-
-
-@pytest.fixture
-def mock_llm_request():
-    """Provide a mock LLM request for testing."""
-    return MockLlmRequest()
-
-
-@pytest.fixture
-def mock_llm_response():
-    """Provide a mock LLM response for testing."""
-    return MockLlmResponse(text="Test response")
-
-
-@pytest.fixture
-def mock_tool():
-    """Provide a mock tool for testing."""
-    return MockTool("test_tool", {"arg1": "value1"})
-
-
-@pytest.fixture
-def mock_tool_context():
-    """Provide a mock tool context for testing."""
-    return MockToolContext()
-
-
-@pytest.fixture
-def test_metrics():
-    """Provide a MetricsCollector instance for collecting test metrics."""
-    return MetricsCollector()
-
-
-@pytest.fixture
-def mock_agent():
-    """Provide a mock DevOps agent for testing."""
-    agent = Mock()
-    agent.name = "test_agent"
-    agent.model = "test_model"
-    agent._state_manager = StateManager()
-    agent._context_manager = Mock()
-    agent._planning_manager = Mock()
-    agent._console = Mock()
-    agent._status_indicator = None
-    agent._actual_llm_token_limit = 100000
-    agent.llm_client = Mock()
-
-    # Mock methods
-    agent._is_retryable_error = Mock(return_value=True)
-    agent._count_tokens = Mock(return_value=100)
-    agent._start_status = Mock()
-    agent._stop_status = Mock()
-
-    return agent
 
 
 @pytest.fixture
@@ -179,6 +82,13 @@ def pytest_configure(config):
     config.addinivalue_line("markers", "e2e: mark test as an end-to-end test")
     config.addinivalue_line("markers", "slow: mark test as slow running")
     config.addinivalue_line("markers", "async_test: mark test as async")
+    config.addinivalue_line("markers", "performance: Performance test")
+    config.addinivalue_line("markers", "stress: Stress test")
+    config.addinivalue_line("markers", "load: Load test")
+    config.addinivalue_line("markers", "foundation: Foundation phase tests")
+    config.addinivalue_line("markers", "core: Core integration phase tests")
+    config.addinivalue_line("markers", "orchestration: Tool orchestration phase tests")
+    config.addinivalue_line("markers", "verification: Performance verification phase tests")
 
 
 def pytest_collection_modifyitems(config, items):  # noqa: ARG001
@@ -215,9 +125,23 @@ def test_config():
 
 # Custom pytest hooks for better test reporting
 def pytest_runtest_setup(item):
-    """Setup hook called before each test."""
+    """Setup for each test run."""
     # Log test start
     logging.getLogger("test").info(f"Starting test: {item.name}")
+
+    # Skip performance tests on slow systems
+    if "performance" in item.keywords:
+        if os.environ.get("SKIP_PERFORMANCE_TESTS", "false").lower() == "true":
+            pytest.skip("Performance tests skipped on slow systems")
+
+    # Skip stress tests unless explicitly requested
+    if "stress" in item.keywords:
+        if os.environ.get("RUN_STRESS_TESTS", "false").lower() != "true":
+            pytest.skip("Stress tests skipped unless explicitly requested")
+
+    # Skip load tests unless explicitly requested
+    if "load" in item.keywords and os.environ.get("RUN_LOAD_TESTS", "false").lower() != "true":
+        pytest.skip("Load tests skipped unless explicitly requested")
 
 
 def pytest_runtest_teardown(item, nextitem):  # noqa: ARG001
