+++ tests/integration/optimization/conftest.py
@@ -1,0 +1,3 @@
+"""
+Fixtures for optimization integration tests.
+"""
