--- tests/integration/test_conftest_example.py
@@ -1,379 +1,0 @@
-"""
-Example integration test demonstrating conftest.py fixture usage.
-
-This test serves as documentation and validation for the integration test
-infrastructure provided by conftest.py.
-"""
-
-import pytest
-
-
-@pytest.mark.integration
-@pytest.mark.foundation
-class TestConftestFixtures:
-    """Test class demonstrating conftest.py fixture usage."""
-
-    def test_basic_fixtures(self, mock_llm_client, mock_session_state, test_workspace):
-        """Test basic fixture usage."""
-        # Test that basic fixtures are available and properly configured
-        assert mock_llm_client is not None
-        assert hasattr(mock_llm_client, "generate")
-
-        assert mock_session_state is not None
-        assert isinstance(mock_session_state, dict)
-        # Check for expected keys in session state
-        expected_keys = [
-            "agent_coordination",
-            "code_review",
-            "code_snippets",
-            "context_state",
-        ]
-        for key in expected_keys:
-            assert key in mock_session_state
-
-        assert test_workspace is not None
-        assert "workspace_dir" in test_workspace
-
-        print("✅ Basic fixtures working correctly")
-
-    def test_context_management_fixtures(
-        self, mock_context_manager, mock_smart_prioritizer, mock_rag_system
-    ):
-        """Test context management fixtures."""
-        # Test context manager
-        mock_context_manager.add_code_snippet("test.py", "print('hello')")
-        mock_context_manager.add_tool_result("test_tool", {"result": "success"})
-
-        context, token_count = mock_context_manager.assemble_context(10000)
-        assert context is not None
-        assert token_count > 0
-        assert len(mock_context_manager.code_snippets) == 1
-        assert len(mock_context_manager.tool_results) == 1
-
-        # Test smart prioritizer
-        snippets = [{"content": "test code", "file_path": "test.py"}]
-        prioritized = mock_smart_prioritizer.prioritize_code_snippets(snippets, "test context")
-        assert len(prioritized) == 1
-        assert hasattr(prioritized[0]["_relevance_score"], "final_score")
-
-        # Test RAG system
-        rag_results = mock_rag_system.query("test query", top_k=3)
-        assert len(rag_results) == 3
-        assert all("content" in result for result in rag_results)
-
-        print("✅ Context management fixtures working correctly")
-
-    def test_agent_fixtures(self, mock_devops_agent, mock_software_engineer_agent, mock_swe_agent):
-        """Test agent fixtures."""
-        agents = [mock_devops_agent, mock_software_engineer_agent, mock_swe_agent]
-
-        for agent in agents:
-            assert agent is not None
-            assert hasattr(agent, "name")
-            assert hasattr(agent, "context_manager")
-            assert hasattr(agent, "llm_client")
-            assert hasattr(agent, "process_message")
-
-        # Test agent names
-        assert mock_devops_agent.name == "devops_agent"
-        assert mock_software_engineer_agent.name == "software_engineer_agent"
-        assert mock_swe_agent.name == "swe_agent"
-
-        print("✅ Agent fixtures working correctly")
-
-    @pytest.mark.asyncio
-    async def test_async_fixtures(self, mock_workflow_engine, mock_tool_orchestrator):
-        """Test async fixture usage."""
-        # Test workflow engine
-        workflow_result = await mock_workflow_engine.execute_workflow(
-            "test_workflow", ["agent1", "agent2"], {"config": "test"}
-        )
-        assert workflow_result["success"] is True
-        assert workflow_result["workflow_type"] == "test_workflow"
-
-        # Test tool orchestrator
-        tool_result = await mock_tool_orchestrator.execute_tool(
-            "test_tool", {"arg1": "value1"}, dependencies=[], tool_id="test_tool_1"
-        )
-        assert tool_result.status == "COMPLETED"
-        assert tool_result.result["success"] is True
-
-        print("✅ Async fixtures working correctly")
-
-    def test_performance_fixtures(self, mock_performance_monitor, mock_resource_monitor):
-        """Test performance monitoring fixtures."""
-        # Test performance monitor
-        mock_performance_monitor.start_monitoring()
-        mock_performance_monitor.record_operation(success=True)
-        mock_performance_monitor.record_operation(success=False)
-
-        metrics = mock_performance_monitor.stop_monitoring()
-        assert hasattr(metrics, "execution_time")
-        assert hasattr(metrics, "peak_memory_mb")
-        assert hasattr(metrics, "operations_per_second")
-
-        # Test resource monitor
-        memory_usage = mock_resource_monitor.get_memory_usage()
-        cpu_usage = mock_resource_monitor.get_cpu_usage()
-        assert isinstance(memory_usage, float)
-        assert isinstance(cpu_usage, float)
-
-        print("✅ Performance fixtures working correctly")
-
-    def test_test_data_fixtures(
-        self, diverse_code_snippets, realistic_tool_results, conversation_history
-    ):
-        """Test data fixtures."""
-        # Test code snippets
-        assert len(diverse_code_snippets) > 0
-        for snippet in diverse_code_snippets:
-            assert "code" in snippet or "content" in snippet  # Support both formats
-            assert "file_path" in snippet
-
-        # Test tool results
-        assert len(realistic_tool_results) > 0
-        for result in realistic_tool_results:
-            assert "tool" in result
-            assert "result" in result
-
-        # Test conversation history
-        assert len(conversation_history) > 0
-        for turn in conversation_history:
-            assert (
-                "user_message" in turn or "agent_message" in turn or "message" in turn
-            )  # Support different formats
-            assert "timestamp" in turn or "tool_calls" in turn  # Support different formats
-
-        print("✅ Test data fixtures working correctly")
-
-    def test_configuration_fixtures(self, test_config, workflow_configs):
-        """Test configuration fixtures."""
-        # Test general config
-        assert "max_execution_time" in test_config
-        assert "token_limit" in test_config
-        assert "performance_thresholds" in test_config
-
-        # Test workflow configs
-        expected_workflows = ["sequential", "parallel", "iterative", "human_in_loop"]
-        for workflow_type in expected_workflows:
-            assert workflow_type in workflow_configs
-            config = workflow_configs[workflow_type]
-            assert "type" in config
-            assert "steps" in config
-            assert "max_retries" in config
-
-        print("✅ Configuration fixtures working correctly")
-
-    def test_phase_specific_fixtures(self, foundation_test_setup, core_integration_setup):
-        """Test phase-specific fixtures."""
-        # Test foundation setup
-        assert "context_manager" in foundation_test_setup
-        assert "agent_pool" in foundation_test_setup
-        assert "workflow_configs" in foundation_test_setup
-
-        # Test core integration setup
-        assert "context_manager" in core_integration_setup
-        assert "smart_prioritizer" in core_integration_setup
-        assert "cross_turn_correlator" in core_integration_setup
-        assert "intelligent_summarizer" in core_integration_setup
-        assert "rag_system" in core_integration_setup
-
-        print("✅ Phase-specific fixtures working correctly")
-
-    def test_metrics_collection(self, test_metrics_collector):
-        """Test metrics collection fixture."""
-        # Record some test metrics
-        test_metrics_collector.record_metric("test_execution_time", 1.5, {"test": "example"})
-        test_metrics_collector.record_metric("test_memory_usage", 100.0, {"unit": "MB"})
-        test_metrics_collector.record_metric("test_operations_count", 50, {"type": "read"})
-
-        # Get metrics
-        metrics = test_metrics_collector.get_metrics()
-        assert len(metrics) == 3
-
-        # Check metric structure
-        for metric in metrics:
-            assert "name" in metric
-            assert "value" in metric
-            assert "timestamp" in metric
-            assert "tags" in metric
-
-        # Get summary
-        summary = test_metrics_collector.get_summary()
-        assert "total_metrics" in summary
-        assert "execution_time" in summary
-        assert "metrics" in summary
-        assert summary["total_metrics"] == 3
-
-        print("✅ Metrics collection working correctly")
-
-    def test_error_simulation_config(self, error_simulation_config):
-        """Test error simulation configuration."""
-        assert "error_types" in error_simulation_config
-        assert "error_probabilities" in error_simulation_config
-        assert "recovery_strategies" in error_simulation_config
-
-        # Check error types
-        expected_errors = ["network_error", "timeout_error", "rate_limit_error"]
-        for error_type in expected_errors:
-            assert error_type in error_simulation_config["error_types"]
-            assert error_type in error_simulation_config["error_probabilities"]
-            assert error_type in error_simulation_config["recovery_strategies"]
-
-        print("✅ Error simulation configuration working correctly")
-
-
-@pytest.mark.integration
-@pytest.mark.core
-class TestParametrizedFixtures:
-    """Test parametrized fixtures from conftest.py."""
-
-    def test_workflow_scenario(self, workflow_scenario):
-        """Test parametrized workflow scenario fixture."""
-        assert "workflow_type" in workflow_scenario
-        assert "agent_count" in workflow_scenario
-
-        # Check valid workflow types
-        valid_types = ["sequential", "parallel", "iterative", "human_in_loop"]
-        assert workflow_scenario["workflow_type"] in valid_types
-        assert isinstance(workflow_scenario["agent_count"], int)
-        assert workflow_scenario["agent_count"] > 0
-
-        print(
-            f"✅ Workflow scenario: {workflow_scenario['workflow_type']} with "
-            f"{workflow_scenario['agent_count']} agents"
-        )
-
-    def test_context_scenario(self, context_scenario):
-        """Test parametrized context scenario fixture."""
-        assert "context_size" in context_scenario
-        assert "token_limit" in context_scenario
-
-        assert isinstance(context_scenario["context_size"], int)
-        assert isinstance(context_scenario["token_limit"], int)
-        assert context_scenario["token_limit"] > context_scenario["context_size"]
-
-        print(
-            f"✅ Context scenario: {context_scenario['context_size']} context size, "
-            f"{context_scenario['token_limit']} token limit"
-        )
-
-
-@pytest.mark.integration
-@pytest.mark.performance
-@pytest.mark.slow
-class TestPerformanceFixtures:
-    """Test performance-related fixtures (marked as slow)."""
-
-    def test_load_test_scenario(self, load_test_scenario):
-        """Test parametrized load test scenario fixture."""
-        assert "concurrent_users" in load_test_scenario
-        assert "operations_per_user" in load_test_scenario
-        assert "expected_min_throughput" in load_test_scenario
-
-        assert isinstance(load_test_scenario["concurrent_users"], int)
-        assert load_test_scenario["concurrent_users"] > 0
-        assert load_test_scenario["operations_per_user"] == 100
-
-        print(
-            f"✅ Load test scenario: {load_test_scenario['concurrent_users']} users, "
-            f"{load_test_scenario['operations_per_user']} ops/user"
-        )
-
-    def test_performance_thresholds(self, test_config):
-        """Test performance threshold configuration."""
-        thresholds = test_config["performance_thresholds"]
-
-        for component in ["context_assembly", "tool_orchestration", "load_testing"]:
-            assert component in thresholds
-            component_thresholds = thresholds[component]
-            assert "max_execution_time" in component_thresholds
-            assert "max_memory_mb" in component_thresholds
-            assert "min_operations_per_second" in component_thresholds
-
-        print("✅ Performance thresholds configured correctly")
-
-
-@pytest.mark.integration
-class TestIntegrationEnvironment:
-    """Test integration environment setup."""
-
-    def test_integration_test_env(self, integration_test_env):
-        """Test integration test environment fixture."""
-        assert "workspace_dir" in integration_test_env
-        assert "test_mode" in integration_test_env
-        assert "mock_llm_client" in integration_test_env
-        assert "session_state" in integration_test_env
-
-        assert integration_test_env["test_mode"] is True
-        assert integration_test_env["mock_llm_client"] is not None
-        assert integration_test_env["session_state"] is not None
-
-        print("✅ Integration test environment working correctly")
-
-    def test_test_reports_dir(self, test_reports_dir):
-        """Test test reports directory fixture."""
-        assert test_reports_dir.exists()
-        assert test_reports_dir.is_dir()
-        assert test_reports_dir.name == "test_reports"
-
-        print("✅ Test reports directory created successfully")
-
-
-# Example of how to use fixtures in realistic test scenarios
-@pytest.mark.integration
-@pytest.mark.foundation
-class TestRealisticIntegrationScenario:
-    """Example of realistic integration test using conftest.py fixtures."""
-
-    @pytest.mark.asyncio
-    async def test_complete_agent_workflow(
-        self,
-        mock_devops_agent,
-        mock_context_manager,
-        mock_workflow_engine,
-        mock_performance_monitor,
-        test_metrics_collector,
-    ):
-        """Test a complete agent workflow using multiple fixtures."""
-        # Start performance monitoring
-        mock_performance_monitor.start_monitoring()
-        test_metrics_collector.record_metric("test_start", 1.0, {"phase": "setup"})
-
-        # Setup context
-        mock_context_manager.add_code_snippet("main.py", "def main(): pass")
-        mock_context_manager.add_tool_result("analyze_code", {"complexity": "low"})
-
-        # Start conversation turn
-        turn_id = mock_context_manager.start_new_turn("Fix the code issues")
-        assert turn_id == 1
-
-        # Process message through agent
-        response = await mock_devops_agent.process_message("Fix the code issues")
-        assert response["success"] is True
-
-        # Execute workflow
-        workflow_result = await mock_workflow_engine.execute_workflow(
-            "fix_issues", [mock_devops_agent], {"priority": "high"}
-        )
-        assert workflow_result["success"] is True
-
-        # Stop performance monitoring
-        performance_metrics = mock_performance_monitor.stop_monitoring()
-        test_metrics_collector.record_metric(
-            "test_execution_time",
-            performance_metrics.execution_time,
-            {"test": "complete_workflow"},
-        )
-
-        # Verify final state
-        context, token_count = mock_context_manager.assemble_context(50000)
-        assert len(context["conversation_history"]) == 1
-        assert len(context["code_snippets"]) == 1
-        assert len(context["tool_results"]) == 1
-
-        print("✅ Complete agent workflow test successful")
-        print(f"   - Turn ID: {turn_id}")
-        print(f"   - Token count: {token_count}")
-        print(f"   - Execution time: {performance_metrics.execution_time:.2f}s")
-        print(f"   - Memory usage: {performance_metrics.peak_memory_mb}MB")
